+++ ChangeLog.md
@@ -3,6 +3,11 @@ Komet Changelog
 This changelog summarizes changes and fixes which are a part of each revision.  For more details on the fixes, refer tracking numbers 
 where provided, and the git commit history.
 
+* 2016/11/18 - 1.40:
+    * Added STAMP time to preferences page
+    * fix a bug on mapping roles
+    * fix a bug with the edit concept button
+
 * 2016/11/18 - 1.39:
     * Fixed refset grid not reloading when new inline concept is shown and refset section is already expanded
     * Fixed refset grid to actually show data